+++ packages/sdk/package.json
@@ -43,9 +43,9 @@
     "@cord.network/did": "workspace:*",
     "@cord.network/identifier": "workspace:*",
     "@cord.network/messaging": "workspace:*",
-    "@cord.network/modules": "workspace:*",
     "@cord.network/network": "workspace:*",
     "@cord.network/statement": "workspace:*",
+    "@cord.network/transform": "workspace:*",
     "@cord.network/types": "workspace:*",
     "@cord.network/utils": "workspace:*"
   }
