+++ packages/statement/package.json
@@ -36,6 +36,7 @@
     "@cord.network/config": "workspace:*",
     "@cord.network/did": "workspace:*",
     "@cord.network/identifier": "workspace:*",
+    "@cord.network/transform": "workspace:*",
     "@cord.network/types": "workspace:*",
     "@cord.network/utils": "workspace:*"
   }
