+++ packages/vc-export/package.json
@@ -35,9 +35,9 @@
   },
   "dependencies": {
     "@cord.network/did": "workspace:*",
-    "@cord.network/modules": "workspace:*",
     "@cord.network/network": "workspace:*",
     "@cord.network/sdk": "workspace:*",
+    "@cord.network/transform": "workspace:*",
     "@cord.network/types": "workspace:*",
     "@cord.network/utils": "workspace:*",
     "@digitalbazaar/vc": "^6.0.1",
